feat: 공용 Modal 템플릿으로 인해 불필요해진 컴포넌트 추가 제거
