--- FrontEnd/src/components/main/Login/LogoutModal/index.ts
@@ -1,3 +1,0 @@
-import LogoutModal from "./LogoutModal";
-
-export default LogoutModal;
