--- FrontEnd/src/components/main/Login/NotLoginModal/index.ts
@@ -1,3 +1,0 @@
-import NotLoginModal from "./NotLoginModal";
-
-export default NotLoginModal;
